Testing Changes

+++ PaymentWalletApplication/src/main/java/com/capg/payment_wallet_application/service/AccountServiceImpl.java
@@ -9,11 +9,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.capg.payment_wallet_application.beans.AccountId;
 import com.capg.payment_wallet_application.beans.BankAccount;
+import com.capg.payment_wallet_application.beans.Customer;
 import com.capg.payment_wallet_application.beans.Wallet;
 import com.capg.payment_wallet_application.dto.BankAccountDTO;
 import com.capg.payment_wallet_application.dto.WalletDTO;
 import com.capg.payment_wallet_application.exception.InvalidInputException;
+import com.capg.payment_wallet_application.exception.WalletNotFoundException;
 import com.capg.payment_wallet_application.repo.IAccountRepository;
+import com.capg.payment_wallet_application.repo.WalletRepo;
 import com.capg.payment_wallet_application.util.AccountUtils;
 import com.capg.payment_wallet_application.util.WalletUtils;
 
@@ -22,6 +25,10 @@ public class AccountServiceImpl implements IAccountService {
 
 	@Autowired
 	IAccountRepository accountRepo;
+	
+	@Autowired
+	WalletRepo walletRepo;
+	
 	final Logger logger = LoggerFactory.getLogger(this.getClass());
 	@Override
 	public WalletDTO addAccount(BankAccount bacc) {
@@ -35,18 +42,35 @@ public class AccountServiceImpl implements IAccountService {
 	@Override
 	public WalletDTO removeAccount(BankAccount bacc) {
 		logger.info("removeAccount() is get intiated");
+		AccountId id = new AccountId(bacc.getAccountNo(),bacc.getIfscCode());
+		BankAccount bankAcc = accountRepo.findById(id).orElse(null);
+		if(bankAcc != null)
+		{
 		Wallet wallet = bacc.getWallet();
 		accountRepo.delete(bacc);
 		logger.info("removeAccount() is get exectued");
 		return WalletUtils.convertToWalletDto(wallet);
+		}
+		else
+		{
+			throw new InvalidInputException("Given Account no is not present");
+		}
 	}
 
 	@Override
 	public List<BankAccountDTO> viewAllAccounts(int walletId) {
 		logger.info("viewAllAccounts() is get intiated");
+		Customer wallet =  walletRepo.findByWalletId(walletId);
+		if(wallet != null)
+		{
 		List<BankAccount> bankAccountList = accountRepo.findByWalletId(walletId);
 		logger.info("viewAllAccounts() is get executed");
 		return AccountUtils.convertToBankAccountDtoList(bankAccountList);
+		}
+		else
+		{
+			throw new WalletNotFoundException("Given wallet is not found");
+		}
 	}
 
 	@Override
